Move util classes into a new subfolder
Add Farmtronics.Utils namespace
